fix: pass dashboardId when validating name
